refactor(youtube): facade import を authoritative owner へ移行する (#3962)
