+++ src/youtube_automation/domains/youtube/video_listing.py
@@ -9,7 +9,7 @@ import logging
 from datetime import datetime, timedelta, timezone
 from typing import TYPE_CHECKING, Dict, List
 
-from youtube_automation.core.adapters.errors import YouTubeAPIError
+from youtube_automation.core.errors import YouTubeAPIError
 
 if TYPE_CHECKING:
     from youtube_automation.domains.analytics.ports import AnalyticsBase  # noqa: F401
